+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Honeycomb OpenTelemetry SDK Changelog
 
 ## v.Next
 
+* feat: expose OpenTelemetry Resource as public property
+
 ## 2.0.0
 
 * maint: Update to OpenTelemetry 2.x.